+++ changelog.md
@@ -1,4 +1,7 @@
 # 1.4.5 changelog
 
+## Additions
+- Add config option for radioactivity's check interval
+
 ## Fixes
 - Fix oxygen system doesn't tick